+++ HazardEditor/src/GUI/Library/Input.h
@@ -16,7 +16,7 @@ namespace WindowElement {
 
 		static void Checkbox(const char* label, bool& value);
 		static void ColorPicker(const char* label, Hazard::Color& color);
-		static void TextureSlot(const char* label, uint8_t& index);
+		static void TextureSlot(const char* label, int16_t& index);
 		static bool Slider(const char* label, float& value, float min = 0.0f, float max = 1000.0f);
 	};
 }
